Remove unsupported types. Added docs.
